Changed type of writerTemplate to Maybe Template.
(Compiled Template, not a String.)

Note that we no longer have the LaTeX, Docbook, or
JATS writers look in the template for strings to
determine whether it is a book or an article, or
whether csquotes is used.

To use csquotes for LaTeX, set `csquotes` in your
variables or metadata. It is no longer sufficient to
put `\usepackage{csquotes}` in your template or header
includes.

To specify a book style, use the `documentclass`
variable or `--top-level-division`.